requirements.txt
python-dotenv
sendgrid==6.0.5
gspread
oauth2client
